+++ server/schedule/hot_cache.go
@@ -108,7 +108,7 @@ func calculateWriteHotThreshold(stores *core.StoresInfo) uint64 {
 	// and we use total written Bytes past storeHeartBeatReportInterval seconds to divide the number of hot Regions
 	// divide 2 because the store reports data about two times than the region record write to rocksdb
 	divisor := float64(statCacheMaxLen) * 2
-	hotRegionThreshold := uint64(float64(stores.TotalBytesWriteRate()) / divisor)
+	hotRegionThreshold := uint64(stores.TotalBytesWriteRate() / divisor)
 
 	if hotRegionThreshold < hotWriteRegionMinFlowRate {
 		hotRegionThreshold = hotWriteRegionMinFlowRate
@@ -121,7 +121,7 @@ func calculateReadHotThreshold(stores *core.StoresInfo) uint64 {
 	// suppose the number of the hot Regions is statLRUMaxLen
 	// and we use total Read Bytes past storeHeartBeatReportInterval seconds to divide the number of hot Regions
 	divisor := float64(statCacheMaxLen)
-	hotRegionThreshold := uint64(float64(stores.TotalBytesReadRate()) / divisor)
+	hotRegionThreshold := uint64(stores.TotalBytesReadRate() / divisor)
 
 	if hotRegionThreshold < hotReadRegionMinFlowRate {
 		hotRegionThreshold = hotReadRegionMinFlowRate
